Fix "event is not defined" error in formlogic.js
While doing the "Creating Your Own Serverless Reminder Service on AWS with Step Functions, API Gateway, Lambda and S3" activity, I found that the buttons in the static site don't work in Firefox. Opening the console shows the following error:
> ReferenceError: event is not defined                    formlogic.js:40:5 

This is happening because the "event" variable isn't being passed to the "sendData" function. The reason it works in Chrome is because it defines a global variable for "event", which is a non-standard kludge for IE compatibility. See [Why is 'event' variable available even when not passed as a parameter?](https://stackoverflow.com/questions/33167092/why-is-event-variable-available-even-when-not-passed-as-a-parameter)